Add opt-in generated fake clients alongside the fake handlers
